feat: added function external key refactor: split file
