+++ cache/cache.go
@@ -3,10 +3,7 @@ package cache
 import (
 	"encoding/json"
 	"fmt"
-	"log"
 	"os"
-	"reflect"
-	"strings"
 	"time"
 
 	"github.com/go-redis/redis"
@@ -159,113 +156,4 @@ func TTL(key string) (float64, error) {
 		return 0, fmt.Errorf("ttl: set duration failed: %s: %s", os.Getenv("REDIS_HOST"), err.Error())
 	}
 	return res, nil
-}
-
-func getKey(data interface{}) (key string, err error) {
-	v := reflect.ValueOf(data)
-
-	// check for nil and pointer dereference
-	if data == nil {
-		return
-	}
-	if v.Kind() == reflect.Ptr {
-		if v.IsNil() {
-			return
-		}
-		v = v.Elem()
-	}
-
-	t := v.Type()
-
-	for i := 0; i < v.NumField(); i++ {
-		field := v.Field(i)
-		fieldT := t.Field(i)
-
-		// check tag exist
-		cacheTag := fieldT.Tag.Get("cache")
-		if len(cacheTag) == 0 {
-			continue
-		}
-		tags := strings.Split(cacheTag, ",")
-
-		// check empty value
-		if tags[0] != "optional" && reflect.DeepEqual(field.Interface(), reflect.Zero(fieldT.Type).Interface()) {
-			return "", fmt.Errorf("redis key: data cannot be empty")
-		}
-
-		// get json tag, else name
-		name := strings.SplitN(fieldT.Tag.Get("json"), ",", 2)[0]
-		if name == "" || name == "-" {
-			name = strings.ToLower(fieldT.Name)
-		}
-
-		// pointer dereference
-		if field.Kind() == reflect.Ptr {
-			field = v.Elem()
-		}
-
-		value := field.Interface()
-
-		// if nested struct
-		if tags[0] != "nodive" {
-			if field.Kind() == reflect.Struct {
-				value, err = getKey(value)
-				if err != nil {
-					return "", err
-				}
-			}
-		}
-
-		key = fmt.Sprintf("%v#%v:%v", key, name, value)
-	}
-	return key, nil
-}
-
-// Key params
-// @data: interface{}
-// @prefixes: ...string
-// return string, error
-func Key(data interface{}, prefixes ...string) (key string) {
-	v := reflect.ValueOf(data)
-	serviceName := os.Getenv("SERVICE_NAME")
-
-	if serviceName == "" {
-		log.Println("redis key: SERVICE_NAME env variable should not be empty")
-		return ""
-	}
-
-	// for non struct based key
-	if data == nil {
-		key = serviceName
-
-		for _, p := range prefixes {
-			key = fmt.Sprintf("%v#%v", key, p)
-		}
-		return key
-	}
-
-	if reflect.DeepEqual(data, reflect.Zero(reflect.TypeOf(data)).Interface()) {
-		log.Println("redis key: data should not be empty")
-		return ""
-	}
-
-	// pointer dereference
-	if v.Kind() == reflect.Ptr {
-		v = v.Elem()
-	}
-
-	key = fmt.Sprintf("%v#%v", serviceName, v.Type().Name())
-
-	for _, p := range prefixes {
-		key = fmt.Sprintf("%v#%v", key, p)
-	}
-
-	dataKey, err := getKey(data)
-	if err != nil {
-		log.Println(err.Error())
-		return ""
-	}
-	key += dataKey
-
-	return key
 }
